+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.3.1 (unreleased)
 
-
+  - Multiple "-c" configuration files can be passed to `statsite`.
+  - If a configuration file doesn't exist or contains a parse
+    error, and error will be raised.
 
 ## 0.3.0 (September 1, 2011)